refactor: pagination state lifted up to lower complexity - PLG-226
