+++ source/javascripts/components/ProductTour/ProductTooltip.tsx
@@ -1,116 +1,42 @@
 import React from "react";
-import { useEffect, useReducer } from "react";
 
 import { Text, Button, Flex, Icon, Link } from "@bitrise/bitkit";
 
 import "./ProductTooltip.scss";
 
-import { State, Action, Tips } from "./types";
+import { Tips } from "./types";
 
 interface ProductTooltipProps {
-	tips: Tips[];
+	tip: Tips;
 	rect: DOMRect | null;
 	onClose: () => void;
-	onChange: (id: string) => void;
 	onButtonClick: (data?: string) => void;
+	onNext: () => void;
+	onPrev: () => void;
+	finished: boolean;
+	selectedIndex: number;
+	total: number;
 }
 
-const initialState: State = {
-	items: [],
-	finished: false
-};
-
-function reducer(state: State, { type, payload }: Action): State {
-	switch (type) {
-		case "reset": {
-			return initialState;
-		}
-		case "init": {
-			const selectedId = (payload as Tips[])[0].id;
-			const selectedIndex = 0;
-			const finished = selectedIndex === (payload as Tips[]).length - 2;
-			return { selectedId, selectedIndex, items: payload as Tips[], finished };
-		}
-
-		case "select": {
-			const found = state.items?.find(item => item.id === payload);
-			if (found) {
-				const selectedIndex = state.items.indexOf(found);
-				const finished = selectedIndex === state.items.length - 1;
-				return { ...state, selectedId: found.id, selectedIndex, finished };
-			}
-			break;
-		}
-		case "next": {
-			const { selectedIndex, items } = state;
-			if (selectedIndex !== undefined) {
-				const nextIndex = selectedIndex + 1;
-				const found = items[nextIndex];
-				if (found) {
-					const finished = selectedIndex === items.length - 2;
-					return {
-						...state,
-						selectedId: found.id,
-						selectedIndex: nextIndex,
-						finished
-					};
-				}
-			}
-			break;
-		}
-		case "prev": {
-			const { selectedIndex, items } = state;
-			if (selectedIndex) {
-				const prevIndex = selectedIndex - 1;
-				const found = items[prevIndex];
-				if (found) {
-					const finished = prevIndex === items.length - 2;
-					return {
-						...state,
-						selectedId: found.id,
-						selectedIndex: prevIndex,
-						finished
-					};
-				}
-			}
-			break;
-		}
+export const ProductTooltip = ({
+	tip,
+	finished,
+	selectedIndex,
+	total,
+	onNext,
+	onPrev,
+	rect,
+	onClose,
+	onButtonClick
+}: ProductTooltipProps) => {
+	if (!tip) {
+		return null;
 	}
-
-	return state;
-}
-
-export const ProductTooltip = ({ tips, rect, onClose, onChange, onButtonClick }: ProductTooltipProps) => {
-	const [{ items, finished, selectedIndex, selectedId }, dispatch] = useReducer(reducer, initialState);
-
-	const tip = items[selectedIndex ?? 0];
-
-	useEffect(() => {
-		dispatch({ type: "init", payload: tips });
-	}, [tips]);
-
-	const onNext = () => {
-		dispatch({ type: "next", payload: undefined });
-	};
-
-	const onPrev = () => {
-		dispatch({ type: "prev", payload: undefined });
-	};
-
-	useEffect(() => {
-		if (selectedId !== undefined) {
-			onChange(selectedId);
-		}
-	}, [selectedId, onChange]);
 
 	const onGotIt = () => {
 		onButtonClick("got it");
 		onClose();
 	};
-
-	if (!tip) {
-		return null;
-	}
 
 	return (
 		<Flex
@@ -162,7 +88,7 @@ export const ProductTooltip = ({ tips, rect, onClose, onChange, onButtonClick }:
 					paddingHorizontal="x3"
 					alignChildrenVertical="middle"
 				>
-					<Text size="2">{selectedIndex !== undefined && <p>{`${selectedIndex + 1}/${items.length}`}</p>}</Text>
+					<Text size="2">{selectedIndex !== undefined && <p>{`${selectedIndex + 1}/${total}`}</p>}</Text>
 				</Flex>
 
 				{finished ? (
