+++ source/javascripts/components/ProductTour/ProductTour.tsx
@@ -1,27 +1,28 @@
 /* eslint-disable @typescript-eslint/camelcase */
 import React, { useEffect } from "react";
-import { useCallback, useState } from "react";
+import { useState } from "react";
 import { Highlighter, useHighlighter } from "./Highlighter";
 import { ProductTooltip } from "./ProductTooltip";
 import { ProductTourProps, Tips } from "./types";
 import { useTrackingFunction } from "./useTrackingFunction";
 import { tips } from "./tips";
 import { useWaitForElements } from "./useWaitForElement";
+import { useProductTour } from "./useProductTour";
 
 export const ProductTourContent = ({ menuIds, currentUser }: ProductTourProps): JSX.Element | null => {
 	const [isOpen, setIsOpen] = useState(true);
 	const [validTips, setValidTips] = useState<Tips[] | null>(null);
-	const [selectedId, setSelectedId] = useState("");
 
-	const onFound = useCallback(() => {
+	const onFound = () => {
 		const filtered = tips.filter(tip => menuIds.includes(tip.id));
 		setValidTips(filtered);
-		setSelectedId(filtered[0].id);
-	}, []);
+	};
 
 	// NOTE: Angular passes the menu items ids it will render,
 	// but we stil need to wait for the templates to render them before we can move on.
 	useWaitForElements(menuIds, onFound);
+
+	const { tip, finished, onNext, onPrev, selectedId, selectedIndex, items } = useProductTour(validTips ?? []);
 
 	const { rect, clipPath, highlighted } = useHighlighter(selectedId);
 
@@ -67,23 +68,25 @@ export const ProductTourContent = ({ menuIds, currentUser }: ProductTourProps): 
 		}
 	}, [isOpen, highlighted, onDisplayTooltip, selectedId]);
 
-	const onChange = useCallback(id => {
-		setSelectedId(id);
-	}, []);
-
 	if (!validTips) {
 		return null;
 	}
 
 	return (
 		<Highlighter isOpen={isOpen} rect={rect!} clipPath={clipPath}>
-			<ProductTooltip
-				onChange={onChange}
-				onClose={onClose}
-				rect={rect}
-				tips={validTips}
-				onButtonClick={onButtonClick}
-			/>
+			{tip && (
+				<ProductTooltip
+					onClose={onClose}
+					onButtonClick={onButtonClick}
+					finished={finished}
+					onNext={onNext}
+					onPrev={onPrev}
+					selectedIndex={selectedIndex!}
+					total={items.length}
+					rect={rect}
+					tip={tip}
+				/>
+			)}
 		</Highlighter>
 	);
 };
