+++ source/javascripts/components/ProductTour/useWaitForElement.ts
@@ -1,13 +1,18 @@
 import { useEffect, useRef } from "react";
 
 export const useWaitForElements = (ids: string[], onFound: (ids: HTMLElement[]) => void) => {
+	const onFoundRef = useRef(onFound);
 	const timeoutRef = useRef<NodeJS.Timeout>();
+	onFoundRef.current = onFound;
+
 	useEffect(() => {
 		const checkIfReady = () => {
 			const found = ids.map(id => document.getElementById(id)).filter(Boolean);
 			if (found.length === ids.length) {
 				clearTimeout(timeoutRef.current);
-				onFound(found as HTMLElement[]);
+				if (onFoundRef.current) {
+					onFoundRef.current(found as HTMLElement[]);
+				}
 				return;
 			}
 			timeoutRef.current = setTimeout(checkIfReady, 250);
@@ -20,5 +25,5 @@ export const useWaitForElements = (ids: string[], onFound: (ids: HTMLElement[]) 
 				clearTimeout(timeoutRef.current);
 			}
 		};
-	}, [ids, onFound]);
+	}, [ids]);
 };
